@@ -8,23 +8,42 @@ using TokenBasedAuth_NetCore.Entities.Trivia;
 
 namespace TokenBasedAuth_NetCore.Context
 {
-    public class UserDbContext: DbContext
-    {
-        public UserDbContext(DbContextOptions options)  : base(options) 
-        {
-                
-        }
-        DbSet<Customer> Customers { get; set; }
-        DbSet<User> Users { get; set; }
 
-    }
+
+
+
     public class TriviaDbContext : DbContext
     {
-        public TriviaDbContext(DbContextOptions options) : base(options)
+        public TriviaDbContext(DbContextOptions<TriviaDbContext> options) : base(options)
         {
 
         }
         DbSet<Category> Categories { get; set; }
 
+    }
+
+    public partial class UserDbContext : DbContext
+    {
+        public UserDbContext(DbContextOptions<UserDbContext> options) : base(options)
+        {
+
+        }
+      
+    }
+
+    public partial class UserDbContext : DbContext
+    {
+  
+        DbSet<Customer> Customers { get; set; }
+        DbSet<User> Users { get; set; }
+
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            base.OnModelCreating(modelBuilder);
+
+    
+
+        }
     }
 }
